+++ examples/button_subscribe.rs
@@ -5,9 +5,10 @@ use futures::future;
 use libtock::buttons;
 use libtock::buttons::ButtonState;
 use libtock::console::Console;
+use libtock::result::TockResult;
 
-// FIXME: Hangs up when buttons are pressed rapidly - problem in console?
-async fn main() {
+// FIXME: Hangs up when buttons are pressed rapidly. Yielding in callback leads to stack overflow.
+async fn main() -> TockResult<()> {
     let mut console = Console::new();
 
     let mut with_callback = buttons::with_callback(|button_num: usize, state| {
@@ -20,13 +21,14 @@ async fn main() {
                 ButtonState::Released => "released",
             }
         )
+        .ok()
         .unwrap();
     });
 
-    let mut buttons = with_callback.init().unwrap();
+    let mut buttons = with_callback.init()?;
 
     for mut button in &mut buttons {
-        button.enable().unwrap();
+        button.enable()?;
     }
 
     future::pending().await