+++ src/electronics/shift_register.rs
@@ -1,4 +1,5 @@
 use crate::gpio::GpioPinWrite;
+use crate::result::TockResult;
 
 pub struct ShiftRegister {
     data_pin: GpioPinWrite,
@@ -19,25 +20,25 @@ impl ShiftRegister {
         }
     }
 
-    pub fn write_bits(&self, values: &[bool]) {
+    pub fn write_bits(&self, values: &[bool]) -> TockResult<()> {
         for i in values {
-            self.push_bit(*i);
+            self.push_bit(*i)?;
         }
-        self.display();
+        self.display()
     }
 
-    fn push_bit(&self, value: bool) {
+    fn push_bit(&self, value: bool) -> TockResult<()> {
         if value {
-            self.data_pin.set_high();
+            self.data_pin.set_high()
         } else {
-            self.data_pin.set_low();
-        }
-        self.clock_pin.set_high();
-        self.clock_pin.set_low();
+            self.data_pin.set_low()
+        }?;
+        self.clock_pin.set_high()?;
+        self.clock_pin.set_low()
     }
 
-    fn display(&self) {
-        self.latch_pin.set_high();
-        self.latch_pin.set_low();
+    fn display(&self) -> TockResult<()> {
+        self.latch_pin.set_high()?;
+        self.latch_pin.set_low()
     }
 }