+++ src/rng.rs
@@ -1,6 +1,8 @@
 use crate::futures;
+use crate::result::TockResult;
 use crate::syscalls;
 use core::cell::Cell;
+use core::mem;
 
 const DRIVER_NUMBER: usize = 0x40001;
 
@@ -16,13 +18,10 @@ mod allow_nr {
     pub const SHARE_BUFFER: usize = 0;
 }
 
-pub async fn fill_buffer(buf: &mut [u8]) -> bool {
+pub async fn fill_buffer(buf: &mut [u8]) -> TockResult<()> {
     let buf_len = buf.len();
 
-    let result = syscalls::allow(DRIVER_NUMBER, allow_nr::SHARE_BUFFER, buf);
-    if result.is_err() {
-        return false;
-    }
+    let shared_memory = syscalls::allow(DRIVER_NUMBER, allow_nr::SHARE_BUFFER, buf)?;
 
     let is_filled = Cell::new(false);
     let mut is_filled_alarm = |_, _, _| is_filled.set(true);
@@ -30,16 +29,14 @@ pub async fn fill_buffer(buf: &mut [u8]) -> bool {
         DRIVER_NUMBER,
         subscribe_nr::BUFFER_FILLED,
         &mut is_filled_alarm,
-    );
-    if subscription.is_err() {
-        return false;
-    }
+    )?;
 
-    let result_code = syscalls::command(DRIVER_NUMBER, command_nr::REQUEST_RNG, buf_len, 0);
-    if result_code < 0 {
-        return false;
-    }
+    syscalls::command(DRIVER_NUMBER, command_nr::REQUEST_RNG, buf_len, 0)?;
 
     futures::wait_until(|| is_filled.get()).await;
-    true
+
+    mem::drop(subscription);
+    mem::drop(shared_memory);
+
+    Ok(())
 }